+++ tests/testthat/test-setMetadata.R
@@ -4,6 +4,8 @@ test_that("setMetadata works", {
                           description = "description",
                           authors = "Gustav Delius",
                           new = "new")
+    # This should not change time_modified
+    expect_identical(params@time_modified, NS_params@time_modified)
     metadata <- getMetadata(params)
     expect_identical(metadata$title, "title")
     expect_identical(metadata$description, "description")